вынес рендер в отдельный модуль
